Validate a file exists before deletion
While uploading vcf.gz files directly to a data library this file may not always exist and attempting to unlink it will cause failure while uploading

+++ lib/galaxy/model/metadata.py
@@ -653,7 +653,8 @@ class FileParameter(MetadataParameter):
                 # directory. Correct.
                 file_name = path_rewriter(file_name)
             mf.update_from_file(file_name)
-            os.unlink(file_name)
+            if os.path.exists(file_name):
+                os.unlink(file_name)
             value = mf.id
         return value
